# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: VegaResume API
  version: 0.2.0
  description: Contract for authentication and resume management.
servers:
  - url: http://localhost:8080
paths:
...
  /api/resumes:
    get:
      operationId: listResumes
      tags:
        - Resume
      summary: List the current user's resumes
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
            maxLength: 80
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ResumeStatus'
        - name: sort
          in: query
          schema:
            $ref: '#/components/schemas/ResumeSort'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            enum: [6, 12, 24]
            default: 6
      responses:
        '200':
          description: Paginated resume summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createResume
      tags:
        - Resume
      summary: Create a blank draft resume
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResumeRequest'
      responses:
        '200':
          description: Created resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/stats:
    get:
      operationId: getResumeStats
      tags:
        - Resume
      summary: Get resume statistics for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user's aggregate resume statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/import:
    post:
      operationId: importResume
      tags:
        - Resume
      summary: Import a VegaResume envelope as a new draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportResumeRequest'
      responses:
        '200':
          description: Imported resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/{resumeId}:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    get:
      operationId: getResume
      tags:
        - Resume
      summary: Get a resume owned by the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resume details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      operationId: updateResume
      tags:
        - Resume
      summary: Update resume metadata or opaque content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResumeRequest'
      responses:
        '200':
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteResume
      tags:
        - Resume
      summary: Delete a resume owned by the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resume deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/{resumeId}/copy:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    post:
      operationId: copyResume
      tags:
        - Resume
      summary: Copy a resume as a new draft
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Copied resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/{resumeId}/import:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    put:
      operationId: replaceResumeImport
      tags:
        - Resume
      summary: Replace an existing resume from a VegaResume envelope
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ImportResumeRequest'
                - type: object
                  required:
                    - expectedRevision
                  properties:
                    expectedRevision:
                      type: integer
                      format: int64
                      minimum: 1
      responses:
        '200':
          description: Replaced resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/{resumeId}/exports:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    post:
      operationId: recordResumeExport
      tags:
        - Resume
      summary: Record a successfully started PDF download
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/resumes/{resumeId}/avatar:
    parameters:
      - $ref: '#/components/parameters/ResumeId'
    get:
      operationId: getResumeAvatar
      tags:
        - Resume
      summary: Get the current resume avatar
      security:
        - bearerAuth: []
      responses:
        '200':
          description: JPEG avatar bytes.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: putResumeAvatar
      tags:
        - Resume
      summary: Replace the current resume avatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteResumeAvatar
      tags:
        - Resume
      summary: Delete the current resume avatar
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated resume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResumeId:
      name: resumeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request, including ErrUsernameFormatInvalid and validation errors.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrUsernameFormatInvalid:
              value:
                code: 101008
                message: Username format is invalid.
                data: null
    Unauthorized:
      description: Authentication failed, token missing, token expired, or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrTokenExpired:
              value:
                code: 101003
                message: Token expired.
                data: null
            ErrTokenInvalid:
              value:
                code: 101004
                message: Token invalid.
                data: null
    Conflict:
      description: Duplicate email or username, including ErrUsernameExists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ErrUsernameExists:
              value:
                code: 101007
                message: Username already exists.
                data: null
...
    InternalServerError:
      description: Server error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    BaseResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          format: int32
          description: 0 for success; non-zero documented error code for failure.
          example: 0
        message:
          type: string
          description: Empty for success; human-readable message for failure.
          example: ''
        data:
          nullable: true
          description: Endpoint-specific payload, or null for errors.
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              example: 100001
            message:
              type: string
              example: Request parameters are invalid.
            data:
              nullable: true
              example: null
    EmptyResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            code:
              type: integer
              format: int32
              enum:
                - 0
            message:
              type: string
              enum:
                - ''
            data:
              nullable: true
              example: null
...
    ResumeStatus:
      type: string
      enum:
        - draft
        - completed
    ResumeSort:
      type: string
      enum:
        - updated_desc
        - updated_asc
        - created_desc
        - title_asc
    ResumeSummary:
      type: object
      required:
        - id
        - title
        - status
        - revision
        - hasAvatar
        - exportCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 80
        status:
          $ref: '#/components/schemas/ResumeStatus'
        revision:
          type: integer
          format: int64
          minimum: 1
        hasAvatar:
          type: boolean
        templateId:
          type: string
          nullable: true
        exportCount:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResumeDetail:
      allOf:
        - $ref: '#/components/schemas/ResumeSummary'
        - type: object
          required:
            - contentVersion
            - content
          properties:
            contentVersion:
              type: integer
              minimum: 1
            content:
              $ref: '#/components/schemas/ResumeContent'
    ResumeStats:
      type: object
      required:
        - total
        - draft
        - completed
        - exported
      properties:
        total:
          type: integer
          minimum: 0
        draft:
          type: integer
          minimum: 0
        completed:
          type: integer
          minimum: 0
        exported:
          type: integer
          format: int64
          minimum: 0
    ResumeListPayload:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResumeSummary'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          enum: [6, 12, 24]
        total:
          type: integer
          minimum: 0
    CreateResumeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
    UpdateResumeRequest:
      type: object
      minProperties: 1
      required:
        - expectedRevision
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 80
        status:
          $ref: '#/components/schemas/ResumeStatus'
        templateId:
          type: string
          nullable: true
        content:
          $ref: '#/components/schemas/ResumeContent'
    ImportResumeRequest:
      type: object
      required:
        - version
        - title
        - content
      properties:
        version:
          type: integer
          enum: [1]
        title:
          type: string
          minLength: 1
          maxLength: 80
        templateId:
          type: string
          nullable: true
          enum:
            - modern-editorial
            - classic-professional
        avatar:
          type: string
          nullable: true
          maxLength: 750000
        content:
          $ref: '#/components/schemas/ResumeContent'
    ResumeContent:
      type: object
      required:
        - profile
        - sections
        - formatting
      additionalProperties: false
      properties:
        profile:
          type: object
          additionalProperties: true
        sections:
          type: array
          maxItems: 64
          items:
            type: object
            additionalProperties: true
        formatting:
          type: object
          additionalProperties: true
    ResumeListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              enum: [0]
            message:
              type: string
              enum: ['']
            data:
              $ref: '#/components/schemas/ResumeListPayload'
    ResumeDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              enum: [0]
            message:
              type: string
              enum: ['']
            data:
              $ref: '#/components/schemas/ResumeDetail'
    ResumeStatsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            code:
              type: integer
              format: int32
              enum: [0]
            message:
              type: string
              enum: ['']
            data:
              $ref: '#/components/schemas/ResumeStats'
...